+++ backend/src/main/java/com/glofox/test/backend/controller/exception/ActivityNotSavedException.java
@@ -1,0 +1,8 @@
+package com.glofox.test.backend.controller.exception;
+
+public class ActivityNotSavedException extends RuntimeException {
+
+    public ActivityNotSavedException(String message) {
+        super("Activity was unexpectedly not saved\n%s".formatted(message));
+    }
+}