+++ backend/src/main/java/com/glofox/test/backend/controller/exception/InconsistentCapacityException.java
@@ -1,0 +1,8 @@
+package com.glofox.test.backend.controller.exception;
+
+public class InconsistentCapacityException extends RuntimeException {
+
+    public InconsistentCapacityException(int capacity) {
+        super("Capacity (%d) must be >= 1".formatted(capacity));
+    }
+}