+++ backend/src/main/java/com/glofox/test/backend/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.glofox.test.backend.repository;
+
+import com.glofox.test.backend.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
+}